refactor(settings/bot): 拆分 BotDetailPage 为 4 个子组件
- BotConfigCard: 基本开关配置（dm/group/echo）
- PushStatusCard: 飞书推送配置（级别、响应开关）
- WhitelistCard: 群聊白名单管理
- HistoryTable: 历史消息分页表格
- BotDetailPage: 改为纯容器，只负责任务编排和状态管理

旧 workspace/BotDetailPage.tsx (543行) → 删除，内容拆分到 settings/bot/
